Add initial Postgresql indexer / importer
It uses Buffalo Pop for the ORM (configured using `database.yml`).
Currently, several credentials are hardcoded mostly for local development.
It is not ready for production use yet, but it does work, and it is
able to pick up on / recover unacknowledged records.

There is also a generic `importer` package to handle most of the
protobuf unpacking, and logging. It can be re-used for other
indexers / importers in the future, e.g. MongoDB.